+++ api/src/controllers/person.js
@@ -9,7 +9,6 @@ const crypto = require("crypto");
 const { catchErrors } = require("../errors");
 const Person = require("../models/person");
 const Team = require("../models/team");
-const RelPersonTeam = require("../models/relPersonTeam");
 const encryptedTransaction = require("../utils/encryptedTransaction");
 const { ENCRYPTED_FIELDS_ONLY, STORAGE_DIRECTORY } = require("../config");
 
@@ -101,7 +100,7 @@ router.post(
       if (ENCRYPTED_FIELDS_ONLY) return data;
 
       // Todo: check if assignedTeams is always needed in full encryption mode.
-      return data.map((p) => ({ ...p.toJSON(), assignedTeams: [] }));
+      return data.map((p) => p.toJSON());
     });
     return res.status(status).send({ ok, data, error });
   })
@@ -126,24 +125,7 @@ router.post(
       if (ENCRYPTED_FIELDS_ONLY) return data;
 
       // Todo: check if assignedTeams is always needed in full encryption mode.
-      if (req.body.hasOwnProperty("assignedTeams")) {
-        await RelPersonTeam.bulkCreate(
-          req.body.assignedTeams.map((teamId) => ({ person: data._id, team: teamId })),
-          { transaction: tx }
-        );
-      }
-
-      // Todo: check if relTeamPerson is always needed in full encryption mode.
-      const relTeamPerson = await RelPersonTeam.findAll({
-        where: {
-          person: data._id,
-        },
-      });
-
-      return {
-        ...data.toJSON(),
-        assignedTeams: relTeamPerson.map((rel) => rel.team),
-      };
+      return data.toJSON();
     });
     return res.status(status).send({ ok, data, error });
   })
@@ -193,20 +175,11 @@ router.get(
 
     // Todo: check if relTeamPerson is always needed in full encryption mode.
     const teams = await Team.findAll(query);
-    const relTeamPersons = await RelPersonTeam.findAll({
-      where: {
-        team: { [Op.in]: teams.map((t) => t._id) },
-      },
-    });
 
     return res.status(200).send({
       ok: true,
       hasMore: data.length === limit,
-      data: data.map((person) => ({
-        ...person.toJSON(),
-        // Todo: check if assignedTeams is always needed in full encryption mode.
-        assignedTeams: relTeamPersons.filter((rel) => rel.person === person._id).map((rel) => rel.team),
-      })),
+      data: data.map((person) => person.toJSON()),
       total,
     });
   })
@@ -247,25 +220,7 @@ router.put(
 
       if (ENCRYPTED_FIELDS_ONLY) return person;
 
-      // Todo: check if relTeamPerson is always needed in full encryption mode.
-      if (req.body.hasOwnProperty("assignedTeams")) {
-        await RelPersonTeam.destroy({ where: { person: req.params._id }, transaction: tx });
-        await RelPersonTeam.bulkCreate(
-          req.body.assignedTeams.map((teamId) => ({ person: req.params._id, team: teamId })),
-          { transaction: tx }
-        );
-      }
-      const relTeamPerson = await RelPersonTeam.findAll({
-        where: {
-          person: person._id,
-        },
-      });
-
-      return {
-        ...newPerson.toJSON(),
-        // Todo: check if assignedTeams is always needed in full encryption mode.
-        assignedTeams: relTeamPerson.map((rel) => rel.team),
-      };
+      return newPerson.toJSON();
     });
 
     res.status(status).send({ ok, data, error });