+++ app/src/recoil/persons.js
@@ -126,6 +126,7 @@ export const addressDetailsFixedFields = [
   'Logement accompagné',
   'Urgence',
   'Insertion',
+  'Hôtel',
 ];
 
 export const addressDetails = [...addressDetailsFixedFields, 'Autre'];