+++ app/src/scenes/Structures/Structure.js
@@ -110,7 +110,7 @@ const Structure = ({ navigation, route }) => {
       return false;
     }
     if (response.ok) {
-      Alert.alert('Structure mise-à-jour !');
+      Alert.alert('Structure mise à jour !');
       setStructures((structures) =>
         structures.map((s) => {
           if (s._id === response.data._id) return response.data;
@@ -289,7 +289,7 @@ const Structure = ({ navigation, route }) => {
         <ButtonsContainer>
           <ButtonDelete onPress={onDeleteRequest} />
           <Button
-            caption={editable ? 'Mettre-à-jour' : 'Modifier'}
+            caption={editable ? 'Mettre à jour' : 'Modifier'}
             onPress={editable ? onUpdateStructure : onEdit}
             disabled={editable ? isUpdateDisabled : false}
             loading={updating}